universal7420: Get rid of all Exynos power HAL stuff
The HIDL implementation of it that all of the stuff removed in this
commit was for has been dropped with LineageOS 19, and bringing it
back requires reverting a few commits, and an AIDL implementation
of it hasn't been written yet.

Now, FB notifier callbacks in the touchscreen and touchkey drivers
are taking care of the only functionality that the Exynos power HAL
was ever useful for, so just get rid of everything for it.