[WIP] universal7420: build stuff from BSP
Our BSP Is incomplete however. HWC is not fully functioning as the BSP needs to be patched to be able to use the internal MPPs. However, it works enough. Once this is fixed up it is easy to support newer versions of android.

whenever i can learn to reverse engineer the blobs or someone helps this will continue to be a problem for us